requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
Pillow==8.3.2

